more type width dependency fixes
git-svn-id: https://gambatte.svn.sourceforge.net/svnroot/gambatte@111 9dfb2916-2d38-0410-aef4-c5fe6c9ffc24

+++ libgambatte/src/video/basic_add_event.h
@@ -15,16 +15,17 @@
  *   version 2 along with this program; if not, write to the               *
  *   Free Software Foundation, Inc.,                                       *
  *   59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.             *
-***************************************************************************/
+ ***************************************************************************/
 #ifndef BASIC_ADD_EVENT_H
 #define BASIC_ADD_EVENT_H
 
+#include <stdint.h>
 #include "ly_counter.h"
 #include "video_event_comparer.h"
 #include "../event_queue.h"
 
 template<class T>
-static inline void addEvent(T &event, const LyCounter &lyCounter, const unsigned cycleCounter, event_queue<VideoEvent*,VideoEventComparer> &queue) {
+static inline void addEvent(T &event, const LyCounter &lyCounter, const unsigned long cycleCounter, event_queue<VideoEvent*,VideoEventComparer> &queue) {
 	if (event.time() == uint32_t(-1)) {
 		event.schedule(lyCounter, cycleCounter);
 		queue.push(&event);
@@ -32,7 +33,7 @@ static inline void addEvent(T &event, const LyCounter &lyCounter, const unsigned
 }
 
 template<class T>
-static inline void addEvent(T &event, const unsigned data, const LyCounter &lyCounter, const unsigned cycleCounter, event_queue<VideoEvent*,VideoEventComparer> &queue) {
+static inline void addEvent(T &event, const unsigned data, const LyCounter &lyCounter, const unsigned long cycleCounter, event_queue<VideoEvent*,VideoEventComparer> &queue) {
 	if (event.time() == uint32_t(-1)) {
 		event.schedule(data, lyCounter, cycleCounter);
 		queue.push(&event);
@@ -40,7 +41,7 @@ static inline void addEvent(T &event, const unsigned data, const LyCounter &lyCo
 }
 
 template<class T>
-static inline void addEvent(T &event, const unsigned data1, const unsigned data2, const LyCounter &lyCounter, const unsigned cycleCounter, event_queue<VideoEvent*,VideoEventComparer> &queue) {
+static inline void addEvent(T &event, const unsigned data1, const unsigned data2, const LyCounter &lyCounter, const unsigned long cycleCounter, event_queue<VideoEvent*,VideoEventComparer> &queue) {
 	if (event.time() == uint32_t(-1)) {
 		event.schedule(data1, data2, lyCounter, cycleCounter);
 		queue.push(&event);
